requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
python-dotenv
openai
pdfplumber
pymupdf
pillow
faiss-cpu
numpy
